ULON-159 (#830)
* ULON-200

계정이 권한을 가진 사이트만 표시해준다.

* ULON-71

출금계좌 세팅 문제 수정

* ULON-219

출금 가승인/최종승인 기능 추가

* 초기화명 변경되는 문제 수정

초기화명 변경되는 문제 수정

* 코드 변경

코드 변경

* status 다시 롤백

status 다시 롤백

* 수정

* 수정

* 수정

* 수정

수정

* KPX 수정

API HOST 설정은 Global Properties 에 정의 하도록 수정

* 기존 용으로 수정

기존 용으로 수정

* 일부 수정

* 수정

* 수정

* 오류 수정

* api 잘못 수정된 부분 수정

api 잘못 수정된 부분 수정

* 파일 선택시에 업로드 진행하도록 수정

파일 선택시에 업로드 진행하도록 수정

* ULON-235

용량 단위 변경

* ULON-236

오류 수정

* oid 세팅 안되는 문제 수정

oid 세팅 안되는 문제 수정

* 알람 개발중

* ULON-240

nullPoint exception 수정

* ULON-97

직접 입력 항목 추가 아직 퍼블 조정 필요

* ULON-242

task 2와 3이 같은 값을 가지도록 변경

* ULON-244, ULON-246

메모 작성 가능하도록 수정

권한 설정 추가

* ULON-249

출금 최종승인 버튼 비활성화 처리

* ULON-248

리스트 갱신 문제 발생 수정

* ULON-250

ULON-250

* 수정

* 최종 승인후 data_send API 호출

최종 승인후 data_send API 호출

* ULON-260

권한 수정

* ULON-238

favicon 잘못세팅되는 문제 수정

* ULON-265

권한 수정

* 수정

수정

* ULON-233

* ULON-230

소문자로 비교

* ULON-222

상업운전 개시일 수정

* ULON-269

밸류 세팅 오류 수정

* ULON-230

검색 오류 수정

* 잘못 수정한 부분 수정

잘못 수정한 부분 수정

* 오류 수정

* 권한 변경

권한 변경

* apiHost 변경

* 엑셀 업로드 까지 작업

* apiHost

 아직 남아있는 호스트 수정

* ULON-159

알람 메세지 등록/수정/조회 개발

* ULON-159

엑셀업로드시에만 적용버튼 활성화 되게 수정

* 삭제 일부 수정

+++ energyServicePortal_ewp_adv/src/main/webapp/WEB-INF/jsp/esp/setting/alarmSetting.jsp
@@ -126,6 +126,10 @@
 					className: 'btn_type fr mb-10',
 					action: function ( e, dt, node, config ) {
 						register();
+					},
+					attr: {
+						id: 'regist',
+						disabled: true
 					}
 				},
 				{
@@ -145,6 +149,10 @@
 					}
 				}
 			],
+			select: {
+				style: 'multi',
+				selector: 'td:first-child > :checkbox'
+			},
 			initComplete: function(settings, json) {
 				let str = `<div id="btnGroup" class="right-end"><!--
 							--><button type="button" disabled class="btn_type03" onclick="deleteRow()">선택 삭제</button><!--
@@ -300,7 +308,7 @@
 
 	// 알람메세지목록 조회
 	const schAlarmList = function (skip) {
-
+		$('#regist').prop('disabled', true);
 		/* validation check */
 		if ($(':checkbox[name="deviceType"]:checked').length == 0) {
 			alert('설비타입을 한개이상 선택해 주세요.');
@@ -448,6 +456,8 @@
 		let alarmTable = $('#alarmTable').DataTable();
 		alarmTable.clear();
 		alarmTable.rows.add(excelArray).draw();
+
+		$('#regist').prop('disabled', false);
 	}
 
 	//적용
@@ -539,50 +549,19 @@
 		let urls = new Array();
 		let deferreds = new Array();
 
-		checkedArray.forEach(chk => {
-			const setId = chk.dataset.setid;
-			const code = chk.dataset.code;
-			urls.push({
-				url: apiHost + '/alarms/code_sets/' + setId + '/codes/' + code,
-				type: 'delete',
-				dataType: 'json',
-			});
-		});
-
-		//코드 삭제 START
-		urls.forEach(function (url) {
-			let deferred = $.Deferred();
-			deferreds.push(deferred);
-
-			$.ajax(url).done(function (data) {
-				data['url'] = url['url'];
-				(function (deferred) {
-					return deferred.resolve(data);
-				})(deferred);
-			}).fail(function (error) {
-				console.log(error);
-			});
-		});
-
-		$.when.apply($, deferreds).then(function () {
-			let setIdArray = new Array();
-			let selectURL = new Array();
-			for (let index = 0; index < arguments.length; index++) {
-				let data = arguments[index];
-				let url = data.url.split('/');
-				setIdArray.push(url[url.length - 3]);
-			}
-
-			//codeSet 조회
-			setIdArray.forEach(setId => {
-				selectURL.push({
-					url: apiHost + '/alarms/code_sets/' + setId + '/codes',
-					type: 'get',
-					dataType: 'json'
+		if ($('#regist').prop('disabled')) {
+			checkedArray.forEach(chk => {
+				const setId = chk.dataset.setid;
+				const code = chk.dataset.code;
+				urls.push({
+					url: apiHost + '/alarms/code_sets/' + setId + '/codes/' + code,
+					type: 'delete',
+					dataType: 'json',
 				});
 			});
 
-			selectURL.forEach(function (url) {
+			//코드 삭제 START
+			urls.forEach(function (url) {
 				let deferred = $.Deferred();
 				deferreds.push(deferred);
 
@@ -597,28 +576,24 @@
 			});
 
 			$.when.apply($, deferreds).then(function () {
-				let deleteURL = new Array();
+				let setIdArray = new Array();
+				let selectURL = new Array();
 				for (let index = 0; index < arguments.length; index++) {
 					let data = arguments[index];
-
-					if (isEmpty(data.data)) {
-						setIdArray = new Array();
-
-						let url = data.url.split('/');
-						setIdArray.push(url[url.length - 2]);
-					}
+					let url = data.url.split('/');
+					setIdArray.push(url[url.length - 3]);
 				}
 
-				//codeSet 삭제
+				//codeSet 조회
 				setIdArray.forEach(setId => {
-					deleteURL.push({
-						url: apiHost + '/alarms/code_sets/' + setId,
-						type: 'delete',
+					selectURL.push({
+						url: apiHost + '/alarms/code_sets/' + setId + '/codes',
+						type: 'get',
 						dataType: 'json'
 					});
 				});
 
-				deleteURL.forEach(function (url) {
+				selectURL.forEach(function (url) {
 					let deferred = $.Deferred();
 					deferreds.push(deferred);
 
@@ -633,13 +608,52 @@
 				});
 
 				$.when.apply($, deferreds).then(function () {
-					alert('삭제가 완료되었습니다.');
+					let deleteURL = new Array();
+					for (let index = 0; index < arguments.length; index++) {
+						let data = arguments[index];
 
-					schAlarmList();
-					return false;
+						if (isEmpty(data.data)) {
+							setIdArray = new Array();
+
+							let url = data.url.split('/');
+							setIdArray.push(url[url.length - 2]);
+						}
+					}
+
+					//codeSet 삭제
+					setIdArray.forEach(setId => {
+						deleteURL.push({
+							url: apiHost + '/alarms/code_sets/' + setId,
+							type: 'delete',
+							dataType: 'json'
+						});
+					});
+
+					deleteURL.forEach(function (url) {
+						let deferred = $.Deferred();
+						deferreds.push(deferred);
+
+						$.ajax(url).done(function (data) {
+							data['url'] = url['url'];
+							(function (deferred) {
+								return deferred.resolve(data);
+							})(deferred);
+						}).fail(function (error) {
+							console.log(error);
+						});
+					});
+
+					$.when.apply($, deferreds).then(function () {
+						alert('삭제가 완료되었습니다.');
+
+						schAlarmList();
+						return false;
+					});
 				});
 			});
-		});
+		} else {
+			alarmTable.rows('.selected').remove().draw( false );
+		}
 	}
 
 	/**
